backend/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.4.2
pydantic-settings>=2.0.3
python-dotenv>=1.0.0
openai>=1.3.0
azure-ai-textanalytics>=5.3.0
azure-ai-formrecognizer>=3.3.0
aiohttp>=3.9.0
python-multipart>=0.0.6
sqlalchemy>=2.0.0
aiosqlite>=0.17.0
azure-cognitiveservices-vision-computervision>=0.9.0
msrest>=0.7.1
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
email-validator>=1.3.0
